+++ examples/data-sources/nxos_isis_interface/data-source.tf
@@ -1,0 +1,3 @@
+data "nxos_isis_interface" "example" {
+  interface_id = "eth1/10"
+}