+++ examples/resources/nxos_isis_interface/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_isis_interface.example "sys/isis/if-[eth1/10]"