+++ examples/resources/nxos_isis_vrf/resource.tf
@@ -2,6 +2,8 @@ resource "nxos_isis_vrf" "example" {
   instance_name           = "ISIS1"
   name                    = "default"
   admin_state             = "enabled"
+  authentication_check_l1 = false
+  authentication_check_l2 = false
   authentication_key_l1   = ""
   authentication_key_l2   = ""
   authentication_type_l1  = "unknown"
@@ -9,7 +11,7 @@ resource "nxos_isis_vrf" "example" {
   bandwidth_reference     = 400000
   banwidth_reference_unit = "mbps"
   is_type                 = "l2"
-  metric_type             = "transition"
+  metric_type             = "wide"
   mtu                     = 2000
   net                     = "49.0001.0000.0000.3333.00"
   passive_default         = "l12"